Saving in case of necessary revert
Lots of dirty test variables, functions, and funky logic in this one. About to undergo a major function rewrite and feature update. This commit has a semi-workable new function in the DeckConverter module to facilitate the display of the file path after a user selects a file in the DeckConverter GUI.  This function needs a logic update to calculate how much of the file path to display based on the pixel-width of the entry widget and characters of the font. Will be deleting, adding, and rewriting a lot of code before the next commit.